Get top holder list working with my remote node... sorta
ok so the top holder list now works against my remote node, being hosted
at graph.wizwar.net. However i realized that it only works if metamask
is available. Ideally i'd want it to work even if metamask isn't
available...

So my plan is to calculate the top holder list based purely off
transfers, which should be easy enough. but i'll take that as an
opportunity to set up some unit testing also so itll be a bit of work.